Remove log4j2 test config from jar differently (#2548)
Redo #2537 to fix #2527 differently. The solution in #2537 was
insufficient because it removed the log4j2-test.properties configuration
file from the class path in IDEs while running tests there, and then
added it back in for the maven-failsafe-plugin. In doing so, it avoided
putting the file in the jar. This change instead explicitly excludes the
file from the jar using the maven-jar-plugin's own excludes
configuration. This ensures the file is on the class path for executing
the unit tests inside the IDE, as well as for maven-failsafe-plugin but
does not place the file in the jar.