Allow UncaughtExceptionHandler to be overridden in client
When an ExecutorService's task or Thread in Accumulo encounter an
unhandled exception the default UncaughtExceptionHandler will be
invoked. An Exception is logged, while an Error is logged and then
the VM is terminated. This change allows clients to override the
default uncaught exception handling logic so that the client VM is
not terminated when an Error happens in a background thread of a
client object.

Closes #2331 and supersedes #2346