Add compareTo method to HostAndPort class (#2551)
The utility class, VerifyTabletAssignements, uses a TreeMap collection
to hold HostAndPort values. This requires the HostAndPort class to
implement the Comparable interface. This ticket adds the compareTo
method to HostAndPort. Given that the utility is seldom used, the
compareTo uses a simple String compare on the Host portion and an 
integer compare on the Port value.

The new method uses Comparator.nullsFirst to deal with cases where the 
parameter value is null. The HostAndPort value if prevented from being null with an
Objects.requireNonNull statement in the fromString method.

A HostAndPortTest class was added to verify changes.

This was first found when running ErrorProne. The ErrorProne check for
this condition had been disabled due to this issue. The check has now
been re-enabled in the ErrorProne configuration within the POM file.


Co-authored-by: Brian Loss <brianloss@gmail.com>