Remove explicit FutureReturnValueIgnored from ErrorProne
By default, all ErrorProne error conditions are checked. There is no
need to explicitly list FutureReturnValueIgnored as an error pattern.
Only error conditions that you do not wish to check for should be
explicitly listed.

Note that it is the opposite for Warning patterns. With the current
configuration you must list any warning patterns you wish to check.