Added mechanism to cancel FaTE operations from Shell with Manager up (#2467)
This change adds a new Thrift API method to the FateServiceHandler to
cancel a new or submitted FaTE transaction. The API method is called from
FateCommand using the ManagerClient class. This will attempt to set
the status to FAILED_IN_PROGRESS and these FaTE transactions will
not be run by the transaction runner threads and instead will be cleaned up.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>