Set default UncaughtExceptionHandler in client, enable override (#2554)
When an ExecutorService's task or Thread in Accumulo encounter an
unhandled exception the default UncaughtExceptionHandler will be
invoked. An Exception is logged, while an Error is logged and then
the VM is terminated. This change sets the UncaughtExceptionHandler
to one that logs the exception/error in the client and does not terminate
the client VM. It also allows clients to supply their own 
UncaughtExceptionHandler implementation.

Closes #2331 and supersedes #2346