Removed Thrift service hierarchy in preparation for service changes
This change modifies TabletClientService such that it no longer extends
ClientService and ManagerClientService such that it no longer extends
FateService. This is in preparation for splitting TabletClientService
into at least two services for the ScanServer feature. The TabletServer
and Manager processes needed to be modified to start two Thrift Processors
instead of one. This required the use of the TMultiplexedProcessor object
on the server side, which takes one or more Processors and places them in
a map keyed on the Thrift service name. Clients, when communicating with
the TabletServer or Manager, then need to use a TMultiplexedProtocol with
the name of the Thrift service that they want to call.