Fix ThriftTransportPool checker loop
Ensure the checker loop checks for shutdown frequently, even if the
configured maximum age for idle transports is longer. This ensures it is
responsive to shutdowns, terminating quickly when a shutdown is
detected, but it doesn't waste time checking for expired idle
connections too often if the max age is relatively long.