Update ChangeSecret to work as documentation states (#2651)
* Update ChangeSecret to work as documentation states

There are inconsistencies in how ChangeSecret should be run, as well as
exceptions thrown when attempting to execute ChangeSecret.

In two locations, instructions for executing ChangeSecret indicate

     ./bin/accumulo org.apache.accumulo.server.util.ChangeSecret

But within accumulo-site.xml, instructions indicate

     ./bin/accumulo org.apache.accumulo.server.util.ChangeSecret [oldpasswd] [newpasswd]

Neither one of the two methods work properly.

Running with no parameters throws a NullPointerException. Running with
the indicated parameters presents a usage message with an error
message provided. Additionally, the usage message makes no mention of
the old and new password parameters.

This PR updates site.xml to match the instructions within the rest of
the code, i.e., execute the command with no parameters.

It also updates the method by which the old and new  parameters
are collected. This allowed the `Opts` class to be removed along 
with other unneeded code. 

The use of `password` is replaced with the string `secret` to 
make the intent more clear.

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>