Rely on WADL for dynamic REST endpoint listing
This removes the http page that shows a static list of hard-coded REST
endpoints added in #2639 and #2649. Instead, this adds the missing
dependency to ensure the WADL endpoint at /rest/application.wadl is
generated properly, so that developers can get an accurate description
of not only the REST endpoint locations, but also their parameters and
return types. Furthermore, since this is using the WADL standards-based
approach, it will be more useful to web-developer tooling to help them
write web views using these endpoints.