Cached last 64 successful ZKSecurityTool.checkCryptPass password checks (#2707)
Apache Commons Codec Crypt.crypt() can be expensive as it creates
numerous MessageDigest objects. Under heavy load the creation of
these MessageDigest objects becomes an issue as there is a synchronization
point in the call stack for MessageDigest.getInstance. As an optimization,
we temporarily cache the last 64 unique inputs to the
ZKSecurityTool.checkCryptPass method that result in a positive return value.

Related to #2700


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>