Wait for server process to notice property changes in MiniAccumuloClusterTest
MiniAccumuloClusterTest.testPerTableClasspath is intermittently failing with
a MutationsRejectedException when the BatchWriter is closed because the TabletServer
is not seeing property changes made in the test right before the BatchWriter is
created. This change introduces a small sleep in an attempt to allow the
TabletServer to react to the configuration changes before the BatchWriter is used
in the test thread.