Move MiniAccumuloClusterTest.testPerTableClasspath to new test class (#2794)
MiniAccumuloClusterTest.testPerTableClasspath is intermittently failing with
a MutationsRejectedException when the BatchWriter is closed because the TabletServer
is not seeing property changes made in the test right before the BatchWriter is
created. This change moves this test method to its own test class so that MAC is
configured with the classpath context configuration properties from the start.