Eventually Consistent scans / ScanServer feature (#2665)
This commit builds on the changes added in prior commits
8be98d6, 50b9267f, and 39bc7a0 to create a new server
component that implements TabletHostingServer and uses
the TabletScanClientService Thrift API to serve client
scan requests on Tablets outside the TabletServer. To
accomplish this the new server (ScanServer) constructs a
new type of tablet called a SnapshotTablet which is comprised
of the files in the metadata table and not the in-memory
mutations that the TabletServer might contain. The Accumulo
client has been modified to allow the user to set a flag
on scans to make them eventually consistent, meaning that
the user is ok with scanning data that may not be immediately
consistent with the version of the Tablet that is being hosted
by the TabletServer.

This feature is optional and experimental.

Closes #2411
Closes #2671
Closes #2733
Closes #2757


Co-authored-by: Keith Turner <kturner@apache.org>