Add a new "dumeta" command to compute table disk usage
This new command differs from the existing command in that it will scan
the Accumulo metadata table for files to compute the usage vs scanning
HDFS files on the file system.. The scan is also done on the client
side instead of on a tablet server. The new command should be
significantly faster than the existing version to compute disk usage
in large clusters.

This closes #2820