Added property to enable Monitor web server to bind to all interfaces
Currently the Monitor uses either the supplied address (-a arg) or the resolved
address for the hostname for advertising it's location in ZooKeeper and the
bind address for the embedded Jetty web server. This commit introduces a new
property that will set the bind address to "0.0.0.0" for the Jetty web server
but will leave the Monitor's ZooKeeper address unchanged. This change is being
made because of some constraints in containerized environments