Added servers per host to cluster.yaml
Added tservers_per_host and sservers_per_host to cluster.yaml file. This file is parsed
and transformed into environment variables that are read in by the accumulo-cluster script.
This change adds the environment variables NUM_TSERVERS and NUM_SSERVERS to the environment
with a default value of 1 if not specified in cluster.yaml and in a way that the value is
in cluster.yaml is overridden if the environment variable already exists in the environment