Serialize location in BadLocationStateException (#2911)
To ensure that the location is correctly serialized in this exception,
do not mark it transient, but convert it to a byte array, since Text is
not Serializable.

This addresses the warning in #2776 without assuming the exception will
never be serialized, which was made in #2897. Instead, this change
avoids errors if it ever were to be serialized and the location were
silently changed to a null value because the field was transient.