Replace du command with new metadata version
This commit updates the "du" command to use the new version that will
scan the metadata table and provide an estimated value. The old api has
been deprecated in favor of the new version. Comments and documentation
have been updated to make it clear the returned data is an estimate.
Shell tests have also been updated for the change in output.