GarbageCollector table existence sanity check (#2792)
Ensure all tables are checked when removing references to candidates
that are still in use. This is done by getting tableIds from zookeeper at
the start of the reference scan and at the end of the reference scan. 
During the reference scans, the list of tableIDs is tracked.  After all candidates
that are still in use have been removed, there is a one more consistency
check to ensure we looked at all tables.

Closes #1377

Co-authored-by: Michael Wall <mjwall@apache.org>