Update "du" command to compute disk usage by scanning the metadata table (#2900)
This commit changes the source of file size information used to compute
shared usages across tables to come from the metadata table instead of
HDFS which will improve performance. This will be an estimate due to
the fact that metadata table contains estimated sizes depending on
what's happening in the cluster. Running a compaction before running the
command gives the most accurate information.

Closes #2820