Re-add getProperties() to InstanceOperations
This method was removed in #2985 because it returned a Thrift object and
appeared to be return the same properties as getSystemConfiguration().
This commit readds the method back (but removes the Thrift object)
as it only returns the properties set in ZK which is different than
getSystemConfiguration() which also returns defaults.