Missing permission checks #2 (#2994)
Added missing permission checks to ClientServiceHandler Thrift API methods
that retrieve the system/namespace/table properties, versioned properties, or
configuration. The system user will be able to call these methods. To get the
system properties / configuration, the caller will need SystemPermission.SYSTEM.
To get the namespace properties / configuration, the caller will need
NamespacePermission.ALTER_NAMESPACE on the namespace. To get the
table properties / configuration, the caller will need TablePermission.ALTER_TABLE
on the table. One side effect of this change is that when cloning a table, if the user
is not the table creator, then the user will need either TablePermission.ALTER_TABLE
on the source table or NamespacePermission.ALTER_NAMESPACE on the namespace.