Print warning in shell when user does not have permission to get config (#3015)
Modified ConfigCommand to only show the user properties that they have
permission to see. When a user is unable to see the properties at a 
specific level warnings are printed in the shell letting them know what
permission they are missing and that the properties that they are seeing
for the table or namespace may actually be set at a higher level that they
can't see. Override information is still shown, but only for levels that the user
has access to.