Change List to SortedSet in RecoveryLogsIterator
This ticket resolves a bug descibed in #3017 (SortedLogrecoveryTest
failures).

That issue described a situation where a large number of tests within
SortedLogRecoverTest.java were failing within certain environments. It
failed consistently for me as well as on the Apache Jenkins server.

The problem was tracked down to an issue where the 'getFiles' method
within RecoveryLogsIterator was returning a List of logFiles that were
presumed to be in lexigraphical order by file name. But on my particular
instance (and presumably the Jenkins server), the files were not returned
in that order. Therefore when 'validateFirstKey' retrieved the first
file from the provided list, it was not actually checking the first key,
but instead, whatever key was in the first file provided by the list. In
cases where the order was not as assumed, the check failed and an exception
was thrown.

The 'getFiles' method has been updated to return a SortedSet of files,
thereby supplying 'validateFirstKey' with the logFiles in expected order.
This change allowed all tests to pass without any additional code
modification.