Change List to SortedSet in RecoveryLogsIterator (#3021)
This ticket resolves a bug described in #3017 (SortedLogRecoveryTest failures).

That issue described a situation where a large number of tests within SortedLogRecoveryTest.java were failing within certain environments. It failed consistently for me as well as on the Apache Jenkins server.

The problem was tracked down to an issue where the 'getFiles' method within RecoveryLogsIterator was returning a List of logFiles that were presumed to be in lexicographical order by file name. But on my particular instance (and presumably the Jenkins server), the files were not returned in that order. Therefore when 'validateFirstKey' retrieved the first file from the provided list, it was not actually checking the first key, but instead, whatever key was in the first file provided by the list. In
cases where the order was not as assumed, the check failed and an exception was thrown.

The 'getFiles' method has been updated to return a SortedSet of files using a Comparator that compares only the final component of the Path value, i.e., the `part-r-xxxxx.xml` portion of the name. This supplies 'validateFirstKey' with the logFiles in expected order. This change allowed all tests to pass without any additional code modification.

Closes #3017