Make use of GitHub Actions environment files (#3024)
set-output was previously used to create the IT matrix, but it is now
deprecated. So, this change makes use of the environment files instead.

See https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#environment-files

Note: the emitted key=value should not quote the value, or else it will be
parsed as a String instead of a JSON type. This took me so long to figure
out, so I'm sharing it here. Also, if you change the name of the output
property, make sure you update the mapping of the step output to the job
output to use the new name.