Move o.a.a.fate to o.a.a.core.fate
Relocate classes from org.apache.accumulo.fate into an appropriate
org.apache.accumulo.core subpackage, in order to ensure a consistent
naming scheme and to ensure all packages in the accumulo-core.jar fall
under a common package that is unique to the jar. This ensures that we
are prepared for any future exports for the JPMS system, and helps us
locate the correct jar given a class name in a stack trace.

This is being done for 2.1, because fate already has breaking changes
from earlier versions, and upgrading to 2.1 already requires no
outstanding fate operations.

* Relocate UtilWaitThread and Retry classes from
  org.apache.accumulo.fate.util to org.apache.accumulo.core.util
* Relocate org.apache.accumulo.fate and its remaining subpackages into
  org.apache.accumulo.core.fate and subpackages under that
* Move unit tests also
* Also remove mentions of accumulo-fate jar and the fate module from the
  test resource's accumulo-site.xml for the manager's upgrade tests