Ensure bundled dependencies are listed in one place (#3039)
Ensure the accuracy of our bundled dependencies by using one canonical
location to define which dependencies get bundled into the binary
tarball. With this change, the assembly descriptor that lists the
bundled dependencies are dynamically computed from the assembly module's
own dependencies. That way, these will never be out of sync, and we only
have one location to update when we modify the dependencies that are
bundled. This ensures we never bundle something we didn't intend, and
aren't missing something we intended to bundle.