Inline unneeded workaround in ClientContext
This removes an unneeded static `memoizeWithExpiration` method that
existed solely to avoid a warning from the modernizer-maven-plugin. This
change does not trigger that warning, and allows us to call
Suppliers.memoizeWithExpiration directly, rather than through a proxy
method. Additionally, this makes each of the three calls to this method
independent of each other, so if it makes sense to tweak the expiration
period for each of them, that can now be done.