Wire up FSDataInputStream.setDropBehind, set to true in FileCompactor (#3079)
Added CachableBlockFile.CachableBuilder.fsPath override method that allows
the caller to specify whether or not setDropBehind should be called on
the FSDataInputStream. Wired up FileOptions and RFileOperations for this
new setting. Modified FileCompactor only to enable this setting. Any scans
that are running on the same Tablet that is compacting will likely have
their blocks cached, so this should not impact scans.

I looked at setting FSDataInputStream.setDropBehind on the FileCompactors
input files after the compaction was completed, in
DatafileManager.bringMajorCompactionOnline, but the way this is wired up
in Hadoop you have to read from the file to make this setting active (see
org.apache.hadoop.hdfs.server.datanode.BlockSender.close().

I believe that the effect of this change is that this will set
POSIX_FADV_DONTNEED on the DataNodes that serve up a copy of each HDFS file
block that is read. This means that DataNodes that serve up different
block replicas may not have the POSIX_FADV_DONTNEED set on the file
descriptor.

The impact of this change is likely inconsequential when looking at a
single file or on a system that is not actively compacting. This change
will have a larger impact on systems that are constantly compacting.


Co-authored-by: Keith Turner <kturner@apache.org>