Use constants for tables instead of string literals
Use existing constants for referencing the table names and IDs for
built-in tables in a few places that previously used hard-coded string
literals. These aren't likely to change, but it makes it easier to find
references of these built-ins when they use the constant, and it's
easier to avoid typos.