Use -source-release.tar.gz vs. -src.tar.gz
Modify the primary project pom.xml file and the assemble pom.xml by
removing the apache-release profiles. Doing so results in the binary and
source distros being located and named to default values and locations.

The binary tarball retains the same name and location as before, whereas
the source tarball will be placed in the same directory structure but now
named accumulo-<VERSION>-source-release.tar.gz, where VERSION is the
current version listed in the pom file.