Remove htrace4 from the binary distribution (#3126)
The htrace4 dependency is not needed for Accumulo. It may or may not be
a dependency of Hadoop, but if it is needed for that, it can be brought
in via the Hadoop lib directory and added to Accumulo's runtime class
path. However, since it is not needed for Accumulo itself, it should not
be redistributed by Accumulo.