Add back in manager PropertyRenamer to ConfigPropertyUpgraderIT to fix test
This fixes the test that was broken by #3106 by adding back in the
master to manager property renamer for just the test. Even though the
old master prefixed properties were removed and upgrade isn't needed
there are no other renamers that currently exist so adding back in the
renamer just for the test will at least verify the functionality still
works.