Refactor `isTablePropertyValid` property method
Refactor the `isTablePropertyValid` method to be more generic.

This change moves the `isValidTablePropertyKey` method calls out
from the `isTablePropertyValid` method and to the calling locations.
This allows for the `isTablePropertyValid` method to be renamed to
`isValidProperty`.

Moving that method allowed for error messages to accurately discribe
an invalid table property error and an invalid value error.

Added 'throws' to method signatures in partial attempt to improve
error visibilites.