Add a last location mode property (#3142)
This adds a property to allow users to change when
the last location is updated in a tablet's metadata.
The last location can be updated on compaction or
on assignment.

In combination with the `manager.startup.tserver.*`
properties, this will allow users to better manage
initial assignments of tablets according to their
preferences, depending on the configured balancer
that respects the last location field.

This fixes #1169

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>