Added logic to validate property value against their types (#3177)
Refactor the `isTablePropertyValid` method to be more generic.
This change moves the `isValidTablePropertyKey` method calls out
from the `isTablePropertyValid` method and to the calling locations.
This allows for the `isTablePropertyValid` method to be renamed to
`isValidProperty`. Moving that method allowed for error messages to
accurately describe an invalid table property error and an invalid
value error. Added 'throws' to method signatures in partial attempt
to improve error visibilities. Adds new `ThriftPropertyException` to
allow for the `IllegalArgumentException` runtime exception to be
caught and displayed to the client.

Closes #3177