Increased number of locks in SynchronousLoadingBlockCache and set them to be fair (#3175)
SynchronousLoadingBlockCache was introduced in ACCUMULO-4641. The code in this class
used to be located in LruBlockCache. In ACCUMULO-4641 the number of load locks was
reduced from 5003 to 2017. There is no comment as to why this was done, but it was
something I noticed while looking at the fairness of the load locks. This commit resets
the default size back to 5003, as it was in 1.10.x. This commit also changes the load
locks to be fair, which allows the different scan threads in the tablet / scan server
to make progress when there is a collision and multiple threads use the same lock.