Address PR feedback
Made logging less verbose in the normal case and more specific in the
recovery case. Modified interface to throw InterruptedException in the
case where the action performed a sleep so that the calling thread
could handle / raise the exception. Change the use of uninterruptible
sleep to Thread.sleep in Action implementations.