Use serialized object instead of byte[] for service locks (#3189)
Renamed ServerServices class to ServiceLockData and modified
ServiceLock to use and return ServiceLockData objects. Prior
to this change the various server processes were creating
their own byte arrays for the service lock data in ZooKeeper.
With this change, the server processes create ServiceLockData
objects and pass that to ServiceLock, which performs the
serialization into ZooKeeeper.