Support writing cluster config parser output to a file
* This commit adds support to provide an optional argument to
ClusterConfigParser to write the output to a file instead of System.out
so that only the output we want is captured and not everything that's
written to standard out.

* This the accumulo-cluster script has also been updated to
to provide the temp file as an output file argument instead of
capturing System.out so that there are not parsing errors by the script if
other output is written to the standard out stream.