Switch back to Retry Object
This commit switches back to using a retry object without an
incremental backoff.
This allows the specified max Wait to always equal the exact amount
of time that the main thread waits for tservers.
It also switches the logging statement to use the retry log instead.
This takes advantage of the Retry object's logging interval.