Fix conditional in LowMemoryDetector (#3241)
The check was executing the action before returning whether or
not the server was running low on memory. For actions that included
a wait, this would cause those processes to wait unnecessarily. This
changes the method such that the action is only executed if the server
is running low on memory

Closes #3240