Added On-Demand table feature
An on-demand table is like an offline table, except that its
tablets will be hosted when a client wants to interact with
them. A table can be put into an on-demand state using the
ondemand shell command or TableOperations.ondemand method.
The accumulo.root and accumulo.metadata tables cannot be put
into an On-Demand state.

Tablets of an on-demand table are hosted when a client fails
to find the tablet location of an on-demand table. When this
occurs the client makes an RPC call that inserts an "ondemand"
column into the Tablet metadata. In the case that an on-demand
tablet needs to be hosted for a client operation, the client
will wait for the tablet to be hosted and the tablet location
to be resolved by the client before proceeding.

The Manager will assign tablets that have an "ondemand"
column in their metadata to TabletServers and will unassign
hosted on-demand tablets when the "ondemand" column does
not exist. When setting an online table to an on-demand
state, all of its tablets will be unloaded. The new Property
MANAGER_TABLET_GROUP_WATCHER_INTERVAL specifies the interval
at which the Manager will look for tablet state changes.
Lower values for this property will reduce the wait time in
the client for an on-demand tablet to be hosted.

TabletServers keep track of the last access time for on-demand
tablets. Periodically the TabletServer will evaluate which
ondemand tablets should be unloaded. The interval for this
evaluation is specified by the new Property
TABLE_ONDEMAND_UNLOADER_INTERVAL. At this interval the
TabletServer will call the OnDemandTabletUnloader class
specified by the Property TABLE_ONDEMAND_UNLOADER, which will
return the set of tablets to unload. Unloading is performed by
removing the "ondemand" column from the tablet metadata, which
will cause the Manager to unassign the tablets. In the case
that the TabletServer is running low on memory it will not
call the OnDemandTabletUnloader, instead unloading the tablet
with the oldest access time.

New metrics are emitted for on-demand tablets. Specifically,
the metric accumulo.tserver.tablets.ondemand.unloaded.lowmem
is incremented when an on-demand tablet is unloaded for low
memory in the TabletServer, and the metric
accumulo.tserver.tablets.ondemand.online is modified when
on-demand tablets are hosted or unloaded.

Finally, a new utility, ListOnlineOnDemandTablets, has been
created to list on-demand tablets that are currently hosted.

Closes #3210 #3211 #3212