Adds support for conditional mutations to Ample
In addition to adding conditional support to Ample, this change also adds new
operation id metadata column.  The operation id is intended to support running
mutually exclusive tablet operations.  The operation id is central to the goal
of using conditional mutations to allow multiple processes to safely modify a
tablet and that is why it is being introduced at the same time as general
conditional mutation support.