Set root location property for paths to contrib files
Replace poorly supported `session.executionRootDirectory` property, and
fragile, hierarchy-dependent `project.parent.basedir` property for paths
to files at the root of the project that are needed by plugin executions
in child modules.

Instead, use the build-helper-maven-plugin's rootlocation goal to
explicitly set a property that points to the root of the multi-module
project to resolve the location of those files consistently.