fixes #3260 by making cached split info a softref (#3261)
The changes in #3249 introduce per tablet caching of split information so that it does not need to be recomputed. If a tablet happens to have large keys and those end up getting cached by the new code in #3249 then it could cause memory pressure or OOMEs in the tablet server.  This commit attempts to avoid those problems by using a softref to store the cached information.