Optimize internal data structures in Authorizations
Instead of creating collections with default initial capacities,
create them with the known capacities. This will reduce the time
taken and amount of memory consumed when populating these objects
during Authorizations object creation when the number of elements
in the Authorizations is larger than the default initial capacity
of the collection data structure. This may also reduce the time
it takes the VisibilityEvaluator to determine if the Authoriations
object contains a token found in a ColumnVisibility.