Resolves server to use for scan in a single place (#3272)
The code for resolving which tserver or sserver to use for a scan was
spread out across multiple methods responsible for executing a scan.
Pulled the code to resolve which server to use into a single place in
the code that executes a scan.

Also introduced a new class to represent the server and server type
(sserver or tserver) used to process a scan.

These changes clean up two problems in the code. First the tablet
server location class was being used to represent a scan server
with a special string placed in the tserver session field. Second
the decision to use a scan server was deeper in the scan code than
error reporting code and the resulted in the need for an odd
instance variable to remember that a scan server was used for error
reporting.  Removing these two problems makes the code easier to
modify and maintain.