Improves scan server resolution in batch scanner code. (#3273)
Refactors the batch scanner code to make two improvements.  First
mapping ranges to scan servers or tservers was moved under one method.
Second the process of mapping ranges to scan servers using the tablet
location cache was improved.  The tablet location cache used to organized
data in a a way that was useful for immediate scans and eventual scans
had to reorganize the data.  A new method was added to the tablet
location cache that passes the raw data to a consumer that can orgranize
it in any way.  This new method is used for scan servers.