fixes #3252 support tablets w/ and w/o location in tablet location cache
Modifies the TabletLocatorCache to support tablets with and without a
location.  Also modified the ScanServerSelector to support optionally
waiting for scan servers when there are none instead of always falling
back to tservers.  The scan severs selector falling back to the tserver
was written with the assumption that tablets are always hosted.  Now
that tablets are loaded onto tservers on demand it may not be desirable
to load the tablets when there are no scan servers.  That is why this
changes also included changes to optionally wait for scan servers when
there are none.  Restarting all scan servers will be a normal
administrative operations, may not want a lot of tablets to all of a
sudden load on tablet servers when the scan servers are restarted.