Add support for storing ranges in DataFileValue
This adds support for storing a collection of ranges associated with a
data file in the Accumulo metadata table. This will allow the ability to
fence off RFiles by ranges in the future. The previous CSV format of the
serialized DataFileValue has been converted to Json and the ranges are
Base64 encoded.