Updates to based on feedback
Removed the DEFAULT hosting goal such that only ALWAYS, NEVER, and
ONDEMAND exist. These values can be set by the user and are stored
in the tablet metadata in a hosting:goal column. When the client
requests hosting of the Tablet, then a hosting:requested column
will be put into the Tablet metadata. The hosting:requested column
will be removed by the OnDemandTabletUnloader when the TabletServer
unloads the Tablet.