Rename TabletLocator to TabletCache
In #3292 the TabletLocator was modified to cache tablets with out
a location. This rename is in response to the more general purpose
of the client side tablet cache.

All changes in the PR were done using IDE refactoring, except for
the addition of one comment. So the code does not really need
review, only the new names.

TabletLocator was rename to TableCache
TabletLocator.locateTablet was rename to findTablet