removes vestigial cache clear in bulk import
While working on #3337 code that cleared the table id cache for no
apparent reason was found.  Investigation indicates that the cache clear
used to happen before a cache read. The cache read was removed, but the
cache clear was left behind.  This commit removes the cache clear.  The
following commit shows the old clear and then read.

https://github.com/apache/accumulo/blob/ff8c875b83880c1fde3562f1280d6e20bc1ac682/server/master/src/main/java/org/apache/accumulo/master/tableOps/bulkVer2/PrepBulkImport.java#L89-L95