improves fetching mulitple extents in ample
Its possible to pass a collection of extents to ample and get the
metadata for those tablets.  However only metadata for the tablets that
exists in the metadata are returned.  Using this method correctly is
tricky because it may in rare cases (like concurrent splits and merges)
only return a subset of what was requested. This commit changes the
ample interface to add handling for these missing extents by adding a
Consumer<KeyExtent> to the interface to which missing extents are
passed.  Now its no longer possible to use this ample functionality
without considering this rare case.  While making this change one
place was found where the missing extents were not being considered.